% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_knn_naive}
\alias{impute_knn_naive}
\title{Impute missing values in a matrix using k-nearest neighbors (k-NN)}
\usage{
impute_knn_naive(obj, miss, k, n_col_miss, method, cores = 1L)
}
\arguments{
\item{obj}{Numeric matrix with missing values represented as NA (NaN).}

\item{miss}{Logical matrix (0/1) indicating missing values (1 = missing).}

\item{k}{Number of nearest neighbors to use for imputation.}

\item{n_col_miss}{Integer vector specifying the count of missing values per column.}

\item{method}{Integer specifying the distance metric: 0 = Euclidean, 1 = Manhattan, 2 = impute.knn.}

\item{cores}{Number of CPU cores to use for parallel processing (default = 1).}
}
\value{
A matrix with imputed values where missing values were present.
}
\description{
This function imputes missing values in a matrix using a k-nearest neighbors
approach based on the specified distance metric. It processes only columns
with missing values, calculating distances to other columns to find the
k-nearest neighbors, and imputes missing values by averaging non-missing
values from these neighbors.
}
